src/test/javascript/spec/app/entities/depot-country-mapping/depot-country-mapping.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { DepotCountryMappingService } from 'app/entities/depot-country-mapping/depot-country-mapping.service';
import { IDepotCountryMapping, DepotCountryMapping } from 'app/shared/model/depot-country-mapping.model';

describe('Service Tests', () => {
  describe('DepotCountryMapping Service', () => {
    let injector: TestBed;
    let service: DepotCountryMappingService;
    let httpMock: HttpTestingController;
    let elemDefault: IDepotCountryMapping;
    let expectedResult: IDepotCountryMapping | IDepotCountryMapping[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(DepotCountryMappingService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new DepotCountryMapping(0, 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a DepotCountryMapping', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new DepotCountryMapping()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a DepotCountryMapping', () => {
        const returnedFromService = Object.assign(
          {
            depotName: 'BBBBBB',
            countryCode: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of DepotCountryMapping', () => {
        const returnedFromService = Object.assign(
          {
            depotName: 'BBBBBB',
            countryCode: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a DepotCountryMapping', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
